Add not keyword and ! operator
This required changes to the parser and adding a method to
the BasicObject class.

I thought the keyword 'not' and the exclamation mark had different
specificity in Ruby during parsing, but I could not find an article
about it nor could I find any differing behavior in irb. So, I made both
'not' and ! behave the same way during parsing.