Add proper singleton class
Instead of storing singleton methods directly on an object, we now build
a proper singleton class, like Ruby does, and store methods on that.

Getting the proper hierarchy for the singleton classes was a bit of a
pain, and I'm still not sure it's all correct. But I think it's at least
close to what Ruby does.

Adding the singleton class also allowed me to move a bunch of methods
off of the Class object and into their proper home, e.g. the ancestors
method was able to move from Class to Module.