Fix nested multiple assignment
I also greatly reduced the complexity in the compiler by punting all the
work to the C runtime. This might not be as fast, but the code is vastly
simpler to grok, which I think is more important at this stage. We can
always optimize this later when we understand it better.