Fix REPL
And add a test!

The neglected REPL has been broken ever since we optimize local
variables. The solution here is a kinda kludgy, but maintains our
variable optimization quite well.

This also exposed that we were not zeroing out a couple of pointers in
nat_alloc, which broke instance variables in the REPL somehow.