Make both gcc and clang happy
I think I finally found the winning combination of argument order and
flags to make both gcc and clang happy to compile everything along with
onigmo statically.

Also, I learned I can pass --with-pic to the configure script to force
position-independent code generation for libonigmo.a. This fixes the
issues with the REPL, yay!