Run GitHub workflow via Docker
I wanted to reduce duplication between the build configuration for
GitHub Actions and the Makefile. Also, I wanted to be able to do
everything that GitHub Actions does, but locally in Docker.

I moved as much of the different steps into tasks in the Makefile as
possible, then made versions of those for running via the Docker
container.

Last, I updated the GitHub workflow to use the Make tasks and Docker.
The workflow runs a bit slower (about 6m vs 4m), but the simplicity and
reduced duplication is worth it IMO.