Build detatched Envs for objects
Our stack-based NatEnvs were being connected via `outer` to the NatEnv
of Class, Module, Block, Proc, etc.

This worked fine as long as we didn't try to traverse up the chain.
Since the Env on the stack is gone, having a pointer to it is just a
recipe for bugs. (☉_☉)