Improve variable allocation
Prior to this commit, we were undisciplined about allocating enough
space on the variable array prior to pushing variables onto it. Now,
we only do the lazy allocation when the REPL is in use, since it is
not possible to know how many variables there will be in the top-level
NatEnv.

This brings us one step closer to being able to use the actual stack
frame for variable storage rather than malloc'd memory.