Put NatEnv back on the heap for now :-(
NatEnv->caller was pointing to stack memory that was getting overwritten.

My mental model for NatEnv is clearly a bit off, so let's go back to
something that works, even though it leaks memory. I'll think on this
for awhile and come back to it when I have a better picture of how to
model Ruby stack frames.

Best case: we make a few adjustments to how NatEnv->caller is stored and
referenced.

Worst case: we leave NatEnv on the heap and teach the the GC how to
collect it.