Put NatEnv back on the C stack
I figured out why NatEnv.caler was sometimes pointing to bad memory...
we were copying it onto objects. They don't need to know the caller, so
we can just set it to NULL. Yay!

I also added an assertion in the GC so hopefully we don't make this
mistake again. :-)