Revert hashmap code back to original
I had originally changed the hash function to use a 32-bit number
instead of 64-bit because I wanted it to fit in our integer object.
This change puts back the original hash function and shifts the 64-bit
int right by 1 to fit. This does lose a bit (haha) of information, but
allows us to retain the orginal hashmap code.