Fix issue with NatHashKey not being relevant for the GC
This allows us remove the hack of disabling the GC during hashmap
operations. Now we're just down to one hack, which is setting NatEnv on
the NatHashKey so that nat_hashmap_hash and nat_hashmap_compare have a
relevant NatEnv to work with.

Later, we'll write our own hashmap library and pass NatEnvs around like
all the rest of Natalie does. :^)