Straight non-idiomatic conversion to C++
This might be completely idiotic, but I wanted to play with C++ a bit
and see if it could help clean up Natalie code and make things a bit
easier to understand.

I'm reading the most excellent "C++ Crash Course" book by Josh
Lospinoso, and an early chapter advocates for using C++ as a "Super C"
-- a C with super powers.

At the very least, the C++ compiler forced me to put `const` qualifiers
on strings where I had been too lazy to do so. So that's a big plus
already!