Use .rb extension rather than .nat
When I started work on Natalie, I thought it might be a significantly
different language than Ruby, mostly inspired by Ruby. But as things
have progressed, I realized that we can be mostly-compatible with Ruby
as well as find exciting new opportunities with an ahead-of-time
compiler!

I want Natalie to be as good a Ruby as possible, and I don't want people
to think I'm building something entirely different.

Also, it looks nice when syntax highlighting works on GitHub. :^)