Fix bug with Array initialization list
Vector.set_size only works as designed when the size is being shrunk. I
should probably rename it, but I don't know what to name it. So I added
an assert so hopefully we don't misuse it again and added a new
set_capacity that is used for growing the underlying array.