Fix backslash weirdness in compiler
This was bugging me for a long time and more recently causing issues
with our efforts at self-hosting. Turns out that String#sub collapses
backslashes when not associated with something to be escaped. Makes some
sense, but was not expected by me. Using a block to String#sub rather
than supplying the replacement string fixes the issue.