Add a new Hashmap<KeyT, T> template
This starts to clean up our usage of David Leeds' hashmap library quite
a bit. More works is needed for ivars, constants, globals, etc.

I also took the opportunity to change how hashmap_put works -- it
now overwrites the data pointer if an entry already exists, which is the
behavior we always want.

A note about std::unordered_map:

I have tried, on a couple occasions, to swap out David Leeds's hashmap
library for the C++ std::unordered_map, which should provide equivalent
functionality. However, every time I've tried it, compilation time slows
dramatically, by approx 2x.

This not only affects compilation time of Natalie overall, but run time
of Natalie, since we are a frontend for gcc/clang. Execution time for
running a Ruby script (which includes compilation) is not good when the
backend process of compilation is doubled.

Thus the reason I have chosen to stick with David Leeds' hashmap
library, with the C++ template wrapper for convenience and to reduce
errors in usage.