Fix Fiber stack memory on OpenBSD
OpenBSD requires the memory used for the stack to marked with MAP_STACK,
otherwise system calls will fail when the kernel checks the stack pointer.

Since each Fiber gets its own stack, we need to use mmap with the
MAP_STACK flag on OpenBSD. Confusingly, there is an identically-named
flag on Linux, but it does something different. So we don't need it
there.