Mark pointers from stack of Fibers too
Note: this is confusing and borderline wrong. Each fiber other than the
"main" fiber has a m_start_of_stack address that is *low*, whereas the
"main" fiber has a m_start_of_stack address that is *high*. This works
ok because our GC knows how to walk the stack either way, but it's very
confusing and will cause bugs down the road.

But, I ran out of brains for this task, so I'm committing this as-is.
Future me will have to take another look at it... :-)