Fix bug creating AssignmentNode on OpenBSD
This might not be OpenBSD's fault -- it might be the particular version
of Clang I'm using there. But in any case, the old code was causing
`operator delete` to be called on an AssignmentNode somewhere. It should
be noted that I was never able to track down *which* AssignmentNode was
being deleted or why.

But creating a temporary variable for `value` here eliminates the call
to `operator delete`. I'm not smart enough to know why yet. :-)