GC: Call Cell() constructor during allocation
I didn't realize this, but the C++17 standard states that the allocation
occurs *before* the arguments to the constructor are evaluated. This
means that we have a half-baked pointer on the heap for a period while
we evaluate more stuff (including code that can call `new`, thus
triggering a collection).

I didn't notice this because GCC, at least with `-std=c++17` set, does
not follow the standard. It evaluates the arguments before allocating
the memory and calling the constructor. (I read online that setting -std
to a newer standard should fix that, but I didn't try it.)

Clang follows the standard, allocating the memory before evaluating
arguments and calling the constructor. Cool.

I believe this would allow me to remove the `AllocationStrategy`
silliness since half-baked objects are no longer a problem.