Fix error in REPL when var and method have same name
[fixes #32]

This has the potential to affect regular code (not run via the REPL),
since I changed how Env::var_get behaves when a variable index is too
high.

This guard with the abort was originally written when our variable
set/get code generation was quite buggy, so I wanted an early abort
to help catch bugs. Now that we're stable in that front, I think it's
fine to just return nil if the variable index is too high.