Fix bugs returning from loops and rescues
Unfortunately, this broke some stuff with the `break` statement,
so I disabled the tests for these things.

While I normally would not commit a known regression, in this case, the
improvement to `return` is so much better, and the work here exposed
some more bugs in `break` that need to be fixed anyway. Everything is in
flux at this point in the project.

I'll work to fix the `break` bugs in a separate commit.