Add new compiler pass for compiling `break`
This commit fixes many issues with both `break` and `return`.

This was a major effort over the course of a couple weeks. I'm not super
happy with the code, but the result is a much more "correct" Ruby
implementation that properly handles break and return in various
contexts.

I'm not thrilled that we essentially have two ways of handling break:
- Value::add_break_flag()
- Env::raise_local_jump_error(...)

Both the compiler and runtime have different code paths for handling
each of these, and it's gross. I *think* it may be possible to get rid
of the 'break flag' concept and only use LocalJumpError, but honestly
I'm quite tired of messing with this stuff for now. I'm just happy the
test suite is passing again and break/return work as I would expect.

Whew. :-)