Fix memory access bug
This fixes a segfault in Natalie when compiled with Clang.

Honestly, I don't know why this is a bug. `m_calling_env` should
only be set for a short while (in Method::call), and only while the
Env is still valid on the stack.

We really need AddressSanitizer up in here. :-)