Remove some numbers in early compiler passes
I added the numbers because pass 1 was re-processing the same sexp again
and I didn't know why. Now I know it was only the :iter sexp, and we can
handle that in the process_iter.

I wish SexpProcessor had a way of marking sexps as processed so they
cannot be processed again...