Fix GC bug in release build
When Natalie is compiled in -O2, the garbage collector was collecting
this `tree` value.

Man I really wish I knew how to fix this in the garbage collector rather
than marking this variable volatile. The fact that the GC cannot see
this intermediate value makes me nervous that there might be other
things it cannot see.

Like, if the `tree` pointer is not on the stack, where is it?? Maybe a
register? Maybe the GC isn't properly scanning registers. Hmmm...