Fix Stub#and_return to actually check arguments
Previously, we checked the passed arguments against the instance
variable `@args` directly. However, when accessing instance variables
inside the method block the instance variables of the object we define
the method on will be used (instead of the instance variables of Stub).

To fix this we move `@args` to a local varible in Stub#and_return.