GC: Fix bug collecting Fiber args
Sometimes we would prematurely collect Fiber arguments because we were
in the middle of the process of switching to the Fiber.

The early return at the top of FiberValue::visit_children was there so we
didn't walk the stack twice. But it also caused us to bail out of visiting
the Fiber arguments and some other stuff. Whoops! :^)