GC: fix bug collecting ManagedVector prematurely in release build
Optimization -O2 was optimizing out the pointer to the vector since it
was no longer needed. The Vector iterator was all that was needed to
iterate over the values in the Vector. Unfortunatley, our GC needs to
continue to see the Vector on the stack or in a register in order to not
collect it and its children.

This fix makes it so the Vector iterator maintains a pointer to the
Vector, so the pointer is still on the stack the whole time. Now our GC
doesn't collect the Vector! :^)