Fix args count when calling a method with a splat
When passing the total count of arguments to #arg_value_by_path we
previously excluded only keyword arguments. However, also splats should
not be counted as actual arguments because there can be none of them
and there is no default.

With the former behavior, calling a method that takes one argument with
a default value and a rest with a splat, would always use the default
value, even if a parameter is passed in the method call.

[#60]