Remove const char* overload of ValuePtr::send
Again (like last commit), I want it to be obvious what is being paid
when calling a method -- a SymbolValue must be looked up.

I have ideas in the future on how to let the compiler help us do such
lookups once and then cache them, but I'm still experimenting. For now,
this is a good step -- it bubbles the work up to the call site so it's
clear what's happening.