Fix bug with lazy enumerators consuming the object
When a lazy enumerator uses an enumerator as an object, it was
previously consuming this enumerator. This means that calling
`to_a` on an enumerator that was already used in a lazy enumerator will
return an empty array. Example:

```ruby
enum = [1,2].lazy
select = enum.select { true }
select.to_a
=> [1,2]
enum.to_a
=> []
```

We fix this by rewinding the enumerator before using it.