Fix bug where GC was collecting hash keys and values too early
TM::Hashmap is not managed by the GC, but HashValue is. If we put a
Value* in a HashValue that is going away and that same pointer in a
TM::Hashmap, the GC may decide to clean up the HashValue and all its
children because it thinks we don't need them any more.