Add Value#equal binding
This is a hack to fix the behavior of `equal?` being called on to
integers. In MRI two integers with the same integer value have the same
object id which is why `1.equal? 1` will evaluate to true. In Natalie
they would have different object ids, which is why this would evaluate
to false. As soon as this is fixed, we can remove this hack.