Start work on exposing the Natalie parser via a C extension to MRI
This brought to light an issue with bootstrapping -- we would need to
have a Ruby parser in order to generate sexp.rb.o. But our goal in
exposing our own parser is to eliminate the need for the ruby_parser
gem. Whoops!

So... I'll be moving sexp.rb back into C++ like it was prior to
73061797e66cb5ec1448daf2698215db15e9fbb0.