Make Array#[] spec compliant
This removes the implementation of ArrayValue#slice and aliases it with
Array#[] (like the docs say).

Array#[] just does the same as slice has done before, except for the
case that only a single value will be retrieved. In this case we do not
need to create a copy of the array we are working with and can just
return the requested value.

[#39]