Don't look through all blocks when allocating new cells
Previously we were looping through each block to find the next one with
a free cell. We had a early return but on some measurements I did the
loop ran on a average 60 iterations.

This patch introduces a list for keeping track which block has free
cells. Each time we want to allocate something we can just use the last
block of this list and do not have to iterate anymore.

Results:

```ruby
(0..100_000).map { |n| n + 1 }
```

The above script runs on master in 2.248s while with this patch it
finishes in 0.228s. (both measured in release mode)