Optimize hash map capacity calculations
Instead of doing pow(2, i) - 1 in a loop, we can instead calculate
the target capacity and then bump it up to the next power of two.
Plus, because capacity is now always a power of two, (it was previously
always 1 less than a power of two) we can use (hash & (capacity - 1))
instead of hash % capacity when converting a hash to an index to further
get a performance boost.