Change how we use ValuePtr in IntegerValue methods
This commit switches the logic from being sorta implicit and magical
into being explicit, thanks to the guarded-ness of ValuePtrs.

Each method has an early "fast" mode of operation if the argument is a
"fast integer". Otherwise, we call unguard() and treat it like a regular
Value*.

This lets us remove a bunch of the mutating is_* methods on ValuePtr,
which were confusing.