Add ArrayObject::ArrayObject(size_t initial_capacity)
Vector::Vector(size_t initial_capacity) allows us to construct a Vector
with a preset capacity, saving us extra work of having to grow the Vector
for the first couple values that get inserted. However, ArrayObject doesn't
currently expose a similar constructor. This is a quick fix to add that.