Use Hashmap to store Allocator blocks instead of Vector
Heap::collect() previously spent a lot of its time checking
whether a candidate block was a member of an allocator (when
collecting roots). This was an O(n) operation, where n is the
number of blocks in the allocator. To speed this up, we can
use a Hashmap as a set to quickly check in O(1). This brings
Heap::collect() down from 27.71% of dynamic instrs to 10.47%
in examples/boardslam.rb and probably benefits other code as
well.

The new code is a little funky and could benefit from a
Hashset class instead of a Hashmap in a future commit.