Optimize RangeObject::include() for fast integers
RangeObject::include() currently works by iterating all values
from begin to end and comparing them to the provided arg. For
fast integers, we can easily compare the argument to the begin
and end values and save a ton of work (O(n) -> O(1), plus we
don't need to hydrate the integers).