Do not leak break flag from a loop
Previously we were leaking the break flag if a loop was processed by the
compiler (I think?).

As I'm not sure if I understood the whole break-situation correctly I'm
also not sure if this change is "good". However, all specs are passing
including the newly added one that failed before.