Use a ManagedVector for Env captured variables
An Env is both a stack frame and a variable scope. When an Env needs to
capture variables, they are stored in the `m_vars` Vector.

For heap-allocated, GC-managed Envs, we have a well-known flow for how
Env-captured variables are intended to be discovered: When the GC walks
all pointers it can find and it encounters an Env* (recognized as a
pointer it manages), the GC calls Env::visit_children(), which walks
the captured variables in the `m_vars` vector.

    Heap-Allocated Env
    ------------------

    | stack |      | GC-managed heap  |     | regular heap  |
    |-------|      |------------------|     |---------------|
    | Env*  |  =>  | Env              |     |               |
    |       |      |  - *m_vars       |  => | Vector<Value> |

                     (^ Env::visit_children() knows to walk m_vars)

However, in many cases, Envs are *not* heap-allocated. They are short-lived
on the stack. When the GC enounters a stack-allocated Env, it does not
recognize it as anything. Any pointers from the Env to GC-managed memory
are recognized as such, which is good, but pointers to elsewhere are not.

    Stack-Allocated Env (before)
    ----------------------------

    | stack     |      | regular heap  |
    |-----------|      |---------------|
    | Env       |      |               |
    |   *m_vars |  =>  | Vector<Value> |

       (^ this pointer is not recognized as anything the GC manages)

The GC is not capable of recognizing pointers to regular (non-managed)
heap, so it skips right over the varibles from the stack-allocated Env.

The Env::visit_children() method is never called on a stack-allocated
Env, so the MarkingVisitor does not know to check all the variables.

By changing this from a SharedPtr<Vector<Value>> to a
ManagedVector<Value>, we allow the GC to spot the variables vector and
visit its children.

    Stack-Allocated Env (after)
    ---------------------------

    | stack     |      | GC-managed heap      |
    |-----------|      |----------------------|
    | Env       |      |                      |
    |   *m_vars |  =>  | ManagedVector<Value> |

       (^ this pointer is recognized and visited by the GC)