Implement new REPL for Natalie
This new REPL should:
- Allow multiline editing (done)
- Allow highlighting (wip)
- Have a navigable history (done)
- Look good
- Profit (?)

This first basic implementation is still a WIP and is only available
when not running in a self-hosted natalie environment. In order to
achieve so I've added 2 new directives nat_ignore_require and
nat_ignore_require_relative which behave usually like a normal require
and require_relative but are ignored by natalie.