Implement a 'canary' pass that optimizes well-know patterns in ruby code
This adds a concept of pass-0, which doesn't 'lower' the AST to the c++ one but
rather transforms a ruby-ast into a 'simplified' ruby-ast. By simplified we do
not mean during compilation time, but rather at runtime, so we can avoid
unnecessary overhead (like allocating an array to get the min/max out of a few
numbers).