Make Integer#** spec compliant
This comes with some fixmes... First of all, we do not support passing a
negative exponent yet, because we do not have Rational.

Also ruby does some weird stuff with bignum exponentiation. There seems
to be some limit when it gets calculated and when not which depends on
the number of bits it needs to be stored. I tried approximating it but
it's probably not even close to MRIs behavior yet.

[#201]