Port Pass0c to the new Compiler and split it up into more passes
This adds a few command line arguments to control the new passes:
```
-f [optimization]          Enable/Disable a specific optimization ...
  nocanary_array_functions Disable unfolding of functions on literal ...
  cf, nocf                 Enable/Disable constant folding [AST], ...
-O1                        Enable basic optimizations (only with -c2)
```
(cropped to fit in 72 characters)
The `[AST]` is a placeholder for now to annotate that these passes
operate on the ast.

The architecture introduced can be easily expanded to toggle more
optimizations in the future.

This also makes the canary-array-pass a bit more error agnostic.