C2: Remove separate and/or instructions in favor of 'if'
and/or in Ruby acts really more like an if statement than a boolean
and/or. I wanted to see if it was possible to implement and/or with the
existing 'if' instruction, and it seems to work.

Doing so did require adding a new 'dup' instruction, but that seems to
be a fairly common instruction for most VMs to have anyway.