C2: start work on begin/rescue
This only handles 'begin' and 'rescue' for now -- no 'else' or 'ensure' yet.

I don't love the MatchExceptionInstruction, but it greatly simplified the
IR that was needed to match an array of exceptions. For now, I think the
tradeoffs justified a dedicated instruction. I reserve the right to
change my mind down the road. :-)