Run C2 and VM tests on MRI
Hosting C2/VM tests on top of Natalie's first-gen compiler was kinda cool,
but tracking down bugs in two levels of compilers proved to be a bit too
difficult for my brain to handle.

The main benefit was flushing out bugs in our Ruby parser anyway, so
running these tests with `--nat-parser` should have much of the same benefit.