C2: use regular C++ local variables when possible
I chose to move the variable scope determination into Pass2 and then
hand that information over to the backend via the individual
instructions... each instruction now has a reference to the "env"
(scope) in which it operates. The C++ backend makes use of this
information.

The VM does not make use of the env info, yet; I will look into
that later. I'm not sure if it would be helpful to the VM or not...