Improve performance of BigInt::divide
Previously ::divide would increment the quotient on each addition of the
divisor. If we have a small divisor and a big dividend this happens a
lot.

Since BigInt operations are rather slow we now increment a long
long, add it to the quotient if the loop ended or if it would exceed it's
max long value.