Delay building an exception backtrace array
Previously, we were building a backtrace array when initializing an
exception object. Sometimes though we do not want to print the backtrace
which makes generating it useless.

This patch tries to delay building the backtrace array, until it will
actually be used.

This should impact a lot of code slightly, because exceptions are
sometimes used for control-flow in the language itself (LocalJumpError
for example).