Fix synthesized float being captured in an ivar
Actually I'm not sure how this fixed the issue... I started to write
more robust guards around auto-hydrating fast integers and floats,
in order to catch the moment a stack-allocated value was being saved
to an instance variable. However, in the process of adding the guard
code, the bug disappeared.