C2: separate Args and MultipleAssignment (no more inheritance)
These two classes started out very similar, and inheritance was useful
then. But as they have diverged, the inheritance was making changes
unpredictable, and more and more methods needed to be overridden anyway.
There are now only a few small methods with duplicate code, which we
could share with a mixin, but I think it's fine to have duplication
where the code is not changing much. :-)