C2: fix bug compiling useless NilObject::the() in Clang
Compiling:

    while false
    end

...was producing this:

    while (condition()->is_truthy()) {
        Value(NilObject::the());
    }

In Clang, this produced the error:

    error: definition or redeclaration of 'the' not allowed inside a function
        Value(NilObject::the());
              ~~~~~~~~~~~^

I believe this is the famous "Most Vexing Parse" issue! I don't know why
GCC knew what I meant, but Clang thought it was a function declaration.
:-(

Anyway, removing useless values during compilation is a worthwhile thing
to do regardless, so it's nice that it fixes the problem in this case.

So, now, the compiled C++ looks like this (the useless nil removed):

    while (condition()->is_truthy()) {
    }