Add possibility to profile Natalie's compiler using StackProf
This patch adds two new options to bin/natalie:

-d transform - Only transform the given ruby code to cpp code
without writing it to a file. This is useful for measuring the time
needed by the compiler to transform ruby code to cpp.

-p compiler - Profiles the compiler using the stackprof gem. When speedscope is
installed, the profile will be opened using speedscope!

Also to profile an application you now have to use `-p app` instead of
just `-p`.