Fix escaping of \ in string output of Regex
The \ character might need some extra escaping, but not when it's used
for a character class. The old code had the following output:

    /\d+/.to_s == "(?-mix:\\\\d+)"

This changed the behaviour: instead of matching one of more decimals,
this regex would match a literal \ followed by one or more d's. This
made it impossible to embed these kind of regexes into another regexp.

This scenario was not defined in the regex specs, but it did
accidentally get caught in the spec for the case statement.