Fix Rational(obj) where obj has a to_r method
This does not have a dependency on the argument being a numeric type. It
should raise a TypeError if the result of the to_r call is not a
Rational.
As a side effect, this required an additional check for the argument
being nil, since `nil.to_r` does return a Rational(0, 1).

There was a slight change needed in the spec, since Rational(String) is
not implemented.