Implement Numeric#i
And remove it from Complex. Even if there is no spec for it, this looks
like the behaviour of MRI.

MRI still enables to rebind the method to call it:

    irb(main):001:0> Numeric.instance_method(:i).bind(Complex(1)).call
    => (0+1i)
    irb(main):002:0> Numeric.instance_method(:i).bind(Complex(1, 2)).call
    => (-2+1i)

This does not work in Natalie, it tries to call `#negative?` on the
Complex value. Given that there is no spec for it, and the MRI behaviour
seems a bit weird given the method `Complex#i` is explicitly removed,
I'm just going to assume this is undefined behaviour.