Test for using send for an undefined method
When this is called from C++, it does try to allocate an ArrayObject on
the stack, the GC expects in on the heap and crashes:
```
..This object (0xXXXXXXXXXXXX) is stack allocated, but you allowed it to
be captured in a Ruby variable or another data structure.
```

The relevant code in the Integer class looks like this:
```c++
bool IntegerObject::eq(Env *env, Value other) {
    // skip
    return other->send(env, "=="_s, { this })->is_truthy();
}
```

This crash does not occur when using `obj.send(:==, 1)` in the test
code.

The NoMethodError is the expected behaviour in the Integer class, that's
how MRI works as well.

We should probably create some custom code for this test, since this
results in a hidden dependency on the implementation of the Integer
class.