Construct Args(TM::Vector) with a reference instead of a copy
There is no need to copy the Vector.

It would be better to use move semantics here, since we simply copy the
size and the data pointer, which means once the argument gets destructed
or its internal value gets reallocated, the Args object has a dangling
pointer. Using move semantics is out of the scope of this branch.