Prevent segfault in String#gsub
It is possible to call String#gsub with only one argument and no block,
in that case the reply should be an enumerator. The code did not handle
that case, and tried to dererence a nullptr into a string.

This change adds the extra case with a NAT_NOT_YET_IMPLEMENTED. It still
halts the program, but at least now it shows what is missing, and it
resolves the security issues of segfaults.