Combine tests in test/runner.rb
This has two advantages.
First of all, this groups the compilation step, so we only have to run
the compiler one. This results in a small performance boost (running the
tests of `spec/library/set` on my machine drop from around 1m35 to 1m20,
so that's around 15% performance increase).
Second, this does continue the run if a single test is failing, and
shows a readable overview at the end of the run.

Because Natalie does not support variables in `load` or `require`, we
had to jump through a few hoops to combine all the arguments into a
single file.