Fix order of instructions when assignment is used
Another way to fix this would be to change the order of instructions
generated for a SendInstruction such that the receiver is always
evaluated and pushed to the stack first. This change proved to be
quite difficult, though, since we have our current operation stack order
is baked into so many places. So I went with the easier fix. :-)