Implement Symbol#encoding
In the worst way possible: currently our Symbols don't have encoding, so
we convert the TM::String into a Ruby String and ask that object for its
encoding. This makes some of the tests pass, but that is more or less
accidentally.

We need to make the encoding a first class citizen of Symbol.