Import specs for OpenSSL::Digest
We had to cheat a bit with the constants: the openssl specs reuse the
data from the Digest module of Ruby (with its own C implementation in
the Ruby source). Since we don't have that, a few lines from these
constants files had to be removed in order to be able to load them.