Disable pipe in Kernel#open
This is meant to be a way to fork a process and return an IO object to
interact with its stdout/stdin, but the current tests result in a file
named '|cat' that was left on your filesystem.
As this feature is a security issue that will be removed in Ruby 4.0
(https://bugs.ruby-lang.org/issues/19630), we probably shouldn't try to
implement this.