Remove keyword swallowing from IO.pipe
This needs more work: the docs of IO describe it as accepting generic IO
creation keywords, the specs only have transcoding options (which we
don't support anyway) that are never mentioned in the IO docs.
For now, simply reject the keyword arguments. This makes it clearer in
our specs that they do not work.