Go back to a single EVAL function in main.cpp
I split this out into `EVAL` and `eval` in
3e7dab3a6f580d2d380105cfc27eb7c277630141 in order to fix top-level
`return`, but it broke the REPL on Linux. :-( The symbol `eval` was not
getting updated with dlopen() and I don't know why. I tried passing a
few different flags to Fiddle::Handle, but it didn't fix the issue. In
the end, I decided to put back our single big EVAL and fix top-level
`return` using a lambda. :-/