Support $LOAD_PATH << something at compile time
...assuming that `something` is available at compile time. For instance:

    $LOAD_PATH << File.expand_path('../../build/prism/lib', __dir__)

^ That sort of code is at the top of a lot of Ruby files, and everything
in it is available at compile time. It's really just a way to manipulate
the `-I load-path` from inside a Ruby script, and it doesn't need to
happen at runtime... at least in this narrow case.

This is a bit on the weird side--combining compiler and VM--I admit,
but I don't think I hate it. It's something that is uniquely Natalie,
and it allows us to support a common Ruby idiom using our VM, but still
at compile time.