Implement Kernel.system (it's complicated)
This is simply a proxy into chaining `Kernel.spawn` and `Process.wait`
with the arguments forwarded. This means this version has the same
limitations as `Kernel.spawn`, most notably it only supports arguments
to the executable as separate arguments to the method, which means
`system("echo foo")` fails. For this reason, the specs have not been
included.

Main rationale for adding this is the nightly spec runner. Currently
pretty much every spec in `core/kernel` fails because `Kernel.system`
does not exist, even though these specs don't need that one. Adding this
method will fix most (or maybe all) of these crashes (currently 96), so
we can see how well we actually do for the Kernel module.