Fix bug in local_variable_or_write transform
The value that might be assigned to the local variable has to be duped.
Otherwise, exiting the current scope of the conditional assignment would
pop of any value of the stack that was there before, potentially erasing
something from the stack that's important. This broke for instance the
following example:

```ruby
case 1
when 1
  text ||= ""
end
```