Update compiler to consume changed Prism nodes
In updating branches to detect new Prism nodes, I realized it was easier
to spot breaking changes when using:

    if node.is_a?(::Prism::SomeNode)
      # do something
    end

...rather than:

    if node.type == :some_node
      # do something
    end

Because the former bit of code raises an exception when nodes are
renamed, whereas the latter will silently fail. So I took the
opportunity to change to the `is_a?` check wherever I touched
any code.