Fix source location for class definition superclass
This is a little silly, adding source location to only the
ConstFindInstruction, but it fixes the line reporting for a bad
superclass. And I'd like to add source location to every instruction at
some point, I just haven't figure out a clean way to do that without
cluttering up pass1. This will work for now.